// src/generated/schema.ts
export interface MediaStatusArgs {
  /** Provide 2 to use new version 2 of sources enum */
  readonly version?: number | null;
}

export interface MediaDescriptionArgs {
  /** Return the string in pre-parsed html instead of markdown */
  readonly asHtml?: boolean | null;
}

export interface MediaSourceArgs {
  /** Provide 2 or 3 to use new version 2 or 3 of sources enum */
  readonly version?: number | null;
}

export interface MediaCharactersArgs {
  readonly sort?: Array<CharacterSort | null> | null;
  readonly role?: CharacterRole | null;
  /** The page */
  readonly page?: number | null;
  /** The amount of entries per page, max 25 */
  readonly perPage?: number | null;
}

export interface MediaStaffArgs {
  readonly sort?: Array<StaffSort | null> | null;
  /** The page */
  readonly page?: number | null;
  /** The amount of entries per page, max 25 */
  readonly perPage?: number | null;
}

export interface MediaStudiosArgs {
  readonly sort?: Array<StudioSort | null> | null;
  readonly isMain?: boolean | null;
}

export interface MediaAiringScheduleArgs {
  /** Filter to episodes that have not yet aired */
  readonly notYetAired?: boolean | null;
  /** The page */
  readonly page?: number | null;
  /** The amount of entries per page, max 25 */
  readonly perPage?: number | null;
}

export interface MediaTrendsArgs {
  readonly sort?: Array<MediaTrendSort | null> | null;
  /** Filter to stats recorded while the media was releasing */
  readonly releasing?: boolean | null;
  /** The page */
  readonly page?: number | null;
  /** The amount of entries per page, max 25 */
  readonly perPage?: number | null;
}

export interface MediaReviewsArgs {
  readonly limit?: number | null;
  readonly sort?: Array<ReviewSort | null> | null;
  /** The page */
  readonly page?: number | null;
  /** The amount of entries per page, max 25 */
  readonly perPage?: number | null;
}

export interface MediaRecommendationsArgs {
  readonly sort?: Array<RecommendationSort | null> | null;
  /** The page */
  readonly page?: number | null;
  /** The amount of entries per page, max 25 */
  readonly perPage?: number | null;
}
// ...
export interface AiringScheduleSelection {
  readonly __typename?: true;
  readonly id?: true;
  readonly airingAt?: true;
  readonly timeUntilAiring?: true;
  readonly episode?: true;
  readonly mediaId?: true;
  readonly media?: MediaSelection;
}

export interface AiringScheduleConnectionSelection {
  readonly __typename?: true;
  readonly edges?: AiringScheduleEdgeSelection;
  readonly nodes?: AiringScheduleSelection;
  readonly pageInfo?: PageInfoSelection;
}
// ...
export interface CharacterConnectionSelection {
  readonly __typename?: true;
  readonly edges?: CharacterEdgeSelection;
  readonly nodes?: CharacterSelection;
  readonly pageInfo?: PageInfoSelection;
}
// ...
export interface FuzzyDateSelection {
  readonly __typename?: true;
  readonly year?: true;
  readonly month?: true;
  readonly day?: true;
}
// ...
export interface MediaSelection {
  readonly __typename?: true;
  readonly id?: true;
  readonly idMal?: true;
  readonly title?: MediaTitleSelection;
  readonly type?: true;
  readonly format?: true;
  readonly status?: { readonly $args: MediaStatusArgs; readonly $select: true };
  readonly description?: { readonly $args: MediaDescriptionArgs; readonly $select: true };
  readonly startDate?: FuzzyDateSelection;
  readonly endDate?: FuzzyDateSelection;
  readonly season?: true;
  readonly seasonYear?: true;
  readonly seasonInt?: true;
  readonly episodes?: true;
  readonly duration?: true;
  readonly chapters?: true;
  readonly volumes?: true;
  readonly countryOfOrigin?: true;
  readonly isLicensed?: true;
  readonly source?: { readonly $args: MediaSourceArgs; readonly $select: true };
  readonly hashtag?: true;
  readonly trailer?: MediaTrailerSelection;
  readonly updatedAt?: true;
  readonly coverImage?: MediaCoverImageSelection;
  readonly bannerImage?: true;
  readonly genres?: true;
  readonly synonyms?: true;
  readonly averageScore?: true;
  readonly meanScore?: true;
  readonly popularity?: true;
  readonly isLocked?: true;
  readonly trending?: true;
  readonly favourites?: true;
  readonly tags?: MediaTagSelection;
  readonly relations?: MediaConnectionSelection;
  readonly characters?: { readonly $args: MediaCharactersArgs; readonly $select: CharacterConnectionSelection };
  readonly staff?: { readonly $args: MediaStaffArgs; readonly $select: StaffConnectionSelection };
  readonly studios?: { readonly $args: MediaStudiosArgs; readonly $select: StudioConnectionSelection };
  readonly isFavourite?: true;
  readonly isFavouriteBlocked?: true;
  readonly isAdult?: true;
  readonly nextAiringEpisode?: AiringScheduleSelection;
  readonly airingSchedule?: { readonly $args: MediaAiringScheduleArgs; readonly $select: AiringScheduleConnectionSelection };
  readonly trends?: { readonly $args: MediaTrendsArgs; readonly $select: MediaTrendConnectionSelection };
  readonly externalLinks?: MediaExternalLinkSelection;
  readonly streamingEpisodes?: MediaStreamingEpisodeSelection;
  readonly rankings?: MediaRankSelection;
  readonly mediaListEntry?: MediaListSelection;
  readonly reviews?: { readonly $args: MediaReviewsArgs; readonly $select: ReviewConnectionSelection };
  readonly recommendations?: { readonly $args: MediaRecommendationsArgs; readonly $select: RecommendationConnectionSelection };
  readonly stats?: MediaStatsSelection;
  readonly siteUrl?: true;
  readonly autoCreateForumThread?: true;
  readonly isRecommendationBlocked?: true;
  readonly isReviewBlocked?: true;
  readonly modNotes?: true;
}

export interface MediaCharacterSelection {
  readonly __typename?: true;
  readonly id?: true;
  readonly role?: true;
  readonly roleNotes?: true;
  readonly dubGroup?: true;
  readonly characterName?: true;
  readonly character?: CharacterSelection;
  readonly voiceActor?: StaffSelection;
}

export interface MediaConnectionSelection {
  readonly __typename?: true;
  readonly edges?: MediaEdgeSelection;
  readonly nodes?: MediaSelection;
  readonly pageInfo?: PageInfoSelection;
}

export interface MediaCoverImageSelection {
  readonly __typename?: true;
  readonly extraLarge?: true;
  readonly large?: true;
  readonly medium?: true;
  readonly color?: true;
}
// ...
export interface MediaExternalLinkSelection {
  readonly __typename?: true;
  readonly id?: true;
  readonly url?: true;
  readonly site?: true;
  readonly siteId?: true;
  readonly type?: true;
  readonly language?: true;
  readonly color?: true;
  readonly icon?: true;
  readonly notes?: true;
  readonly isDisabled?: true;
}

export interface MediaListSelection {
  readonly __typename?: true;
  readonly id?: true;
  readonly userId?: true;
  readonly mediaId?: true;
  readonly status?: true;
  readonly score?: { readonly $args: MediaListScoreArgs; readonly $select: true };
  readonly progress?: true;
  readonly progressVolumes?: true;
  readonly repeat?: true;
  readonly priority?: true;
  readonly private?: true;
  readonly notes?: true;
  readonly hiddenFromStatusLists?: true;
  readonly customLists?: { readonly $args: MediaListCustomListsArgs; readonly $select: true };
  readonly advancedScores?: true;
  readonly startedAt?: FuzzyDateSelection;
  readonly completedAt?: FuzzyDateSelection;
  readonly updatedAt?: true;
  readonly createdAt?: true;
  readonly media?: MediaSelection;
  readonly user?: UserSelection;
}
// ...
export interface MediaRankSelection {
  readonly __typename?: true;
  readonly id?: true;
  readonly rank?: true;
  readonly type?: true;
  readonly format?: true;
  readonly year?: true;
  readonly season?: true;
  readonly allTime?: true;
  readonly context?: true;
}

export interface MediaStatsSelection {
  readonly __typename?: true;
  readonly scoreDistribution?: ScoreDistributionSelection;
  readonly statusDistribution?: StatusDistributionSelection;
  readonly airingProgression?: AiringProgressionSelection;
}

export interface MediaStreamingEpisodeSelection {
  readonly __typename?: true;
  readonly title?: true;
  readonly thumbnail?: true;
  readonly url?: true;
  readonly site?: true;
}
// ...
export interface MediaTagSelection {
  readonly __typename?: true;
  readonly id?: true;
  readonly name?: true;
  readonly description?: true;
  readonly category?: true;
  readonly rank?: true;
  readonly isGeneralSpoiler?: true;
  readonly isMediaSpoiler?: true;
  readonly isAdult?: true;
  readonly userId?: true;
}

export interface MediaTitleSelection {
  readonly __typename?: true;
  readonly romaji?: { readonly $args: MediaTitleRomajiArgs; readonly $select: true };
  readonly english?: { readonly $args: MediaTitleEnglishArgs; readonly $select: true };
  readonly native?: { readonly $args: MediaTitleNativeArgs; readonly $select: true };
  readonly userPreferred?: true;
}

export interface MediaTrailerSelection {
  readonly __typename?: true;
  readonly id?: true;
  readonly site?: true;
  readonly thumbnail?: true;
}

export interface MediaTrendSelection {
  readonly __typename?: true;
  readonly mediaId?: true;
  readonly date?: true;
  readonly trending?: true;
  readonly averageScore?: true;
  readonly popularity?: true;
  readonly inProgress?: true;
  readonly releasing?: true;
  readonly episode?: true;
  readonly media?: MediaSelection;
}

export interface MediaTrendConnectionSelection {
  readonly __typename?: true;
  readonly edges?: MediaTrendEdgeSelection;
  readonly nodes?: MediaTrendSelection;
  readonly pageInfo?: PageInfoSelection;
}
// ...
export interface QuerySelection {
  readonly __typename?: true;
  readonly Page?: { readonly $args: QueryPageArgs; readonly $select: PageSelection };
  readonly Media?: { readonly $args: QueryMediaArgs; readonly $select: MediaSelection };
  readonly MediaTrend?: { readonly $args: QueryMediaTrendArgs; readonly $select: MediaTrendSelection };
  readonly AiringSchedule?: { readonly $args: QueryAiringScheduleArgs; readonly $select: AiringScheduleSelection };
  readonly Character?: { readonly $args: QueryCharacterArgs; readonly $select: CharacterSelection };
  readonly Staff?: { readonly $args: QueryStaffArgs; readonly $select: StaffSelection };
  readonly MediaList?: { readonly $args: QueryMediaListArgs; readonly $select: MediaListSelection };
  readonly MediaListCollection?: { readonly $args: QueryMediaListCollectionArgs; readonly $select: MediaListCollectionSelection };
  readonly GenreCollection?: true;
  readonly MediaTagCollection?: { readonly $args: QueryMediaTagCollectionArgs; readonly $select: MediaTagSelection };
  readonly User?: { readonly $args: QueryUserArgs; readonly $select: UserSelection };
  readonly Viewer?: UserSelection;
  readonly Notification?: { readonly $args: QueryNotificationArgs; readonly $select: NotificationUnionSelection };
  readonly Studio?: { readonly $args: QueryStudioArgs; readonly $select: StudioSelection };
  readonly Review?: { readonly $args: QueryReviewArgs; readonly $select: ReviewSelection };
  readonly Activity?: { readonly $args: QueryActivityArgs; readonly $select: ActivityUnionSelection };
  readonly ActivityReply?: { readonly $args: QueryActivityReplyArgs; readonly $select: ActivityReplySelection };
  readonly Following?: { readonly $args: QueryFollowingArgs; readonly $select: UserSelection };
  readonly Follower?: { readonly $args: QueryFollowerArgs; readonly $select: UserSelection };
  readonly Thread?: { readonly $args: QueryThreadArgs; readonly $select: ThreadSelection };
  readonly ThreadComment?: { readonly $args: QueryThreadCommentArgs; readonly $select: ThreadCommentSelection };
  readonly Recommendation?: { readonly $args: QueryRecommendationArgs; readonly $select: RecommendationSelection };
  readonly Like?: { readonly $args: QueryLikeArgs; readonly $select: UserSelection };
  readonly Markdown?: { readonly $args: QueryMarkdownArgs; readonly $select: ParsedMarkdownSelection };
  readonly AniChartUser?: AniChartUserSelection;
  readonly SiteStatistics?: SiteStatisticsSelection;
  readonly ExternalLinkSourceCollection?: { readonly $args: QueryExternalLinkSourceCollectionArgs; readonly $select: MediaExternalLinkSelection };
}

export interface RecommendationSelection {
  readonly __typename?: true;
  readonly id?: true;
  readonly rating?: true;
  readonly userRating?: true;
  readonly media?: MediaSelection;
  readonly mediaRecommendation?: MediaSelection;
  readonly user?: UserSelection;
}

export interface RecommendationConnectionSelection {
  readonly __typename?: true;
  readonly edges?: RecommendationEdgeSelection;
  readonly nodes?: RecommendationSelection;
  readonly pageInfo?: PageInfoSelection;
}
// ...
export interface ReviewConnectionSelection {
  readonly __typename?: true;
  readonly edges?: ReviewEdgeSelection;
  readonly nodes?: ReviewSelection;
  readonly pageInfo?: PageInfoSelection;
}
// ...
export interface StaffConnectionSelection {
  readonly __typename?: true;
  readonly edges?: StaffEdgeSelection;
  readonly nodes?: StaffSelection;
  readonly pageInfo?: PageInfoSelection;
}
// ...
export interface StudioConnectionSelection {
  readonly __typename?: true;
  readonly edges?: StudioEdgeSelection;
  readonly nodes?: StudioSelection;
  readonly pageInfo?: PageInfoSelection;
}